+++ atlas/code_operations.py
@@ -10,7 +10,7 @@ import subprocess
 import git
 
 from atlas import utilities
-from atlas.config import (ENVIRONMENT, CODE_ROOT, LOCAL_CODE_ROOT)
+from atlas.config import (ENVIRONMENT, CODE_ROOT, WEB_ROOT)
 from atlas.config_servers import (SERVERDEFS)
 
 # Setup a sub-logger. See tasks.py for longer comment.
@@ -69,7 +69,7 @@ def update_symlink_current(item):
     Determine the path for a code item
     """
     code_folder_current = '{0}/{1}/{2}/{2}-current'.format(
-        LOCAL_CODE_ROOT,
+        CODE_ROOT,
         utilities.code_type_directory_name(item['meta']['code_type']),
         item['meta']['name'])
     # Remove symlink if it exists
@@ -84,4 +84,46 @@ def sync_code():
     """
     log.info('Code | Sync')
     hosts = SERVERDEFS[ENVIRONMENT]['webservers'] + SERVERDEFS[ENVIRONMENT]['operations_server']
-    utilities.sync(LOCAL_CODE_ROOT, hosts, CODE_ROOT)
+    # Sync code root
+    utilities.sync(CODE_ROOT, hosts, CODE_ROOT)
+    # Sync static items
+    utilities.sync(WEB_ROOT + '/static', hosts, WEB_ROOT + '/static')
+
+
+def deploy_static(item):
+    """Deploy static asset to the web root
+    
+    Arguments:
+        item {dict} -- Item record for static asset to deploy
+    """
+    log.info('Code | Deploy static')
+    # Create symlink in web root
+    web_root_static_name = WEB_ROOT + '/static/' + item['meta']['name']
+    # Make static directories if needed.
+    if not os.path.isdir(web_root_static_name):
+        os.makedirs(web_root_static_name)
+    # Remove symlink if it exists
+    web_root_path = web_root_static_name + '/' + item['meta']['version']
+    if os.path.islink(web_root_path):
+        os.unlink(web_root_path)
+    os.symlink(utilities.code_path(item), web_root_path)
+
+
+def remove_static(item, other_static_assets=True):
+    """Remove static asset from the web root
+
+    Arguments:
+        item {dict} -- Item record for static asset to deploy
+
+    Keyword Arguments:
+        other_static_assets {bool} -- If false, remove directory for asset name (default: {True})
+    """
+
+    log.info('Code | Remove static')
+    web_root_static_name = WEB_ROOT + '/static/' + item['meta']['name']
+    web_root_path = web_root_static_name + '/' + item['meta']['version']
+    # Remove symlink if it exists
+    if os.path.islink(web_root_path):
+        os.unlink(web_root_path)
+    if not other_static_assets and os.path.isdir(web_root_static_name):
+        os.rmdir(web_root_static_name)